[IMP] real_estate: refine data type for garden_area
- The 'garden_area' field has been updated to a Float type to allow for more precise measurements of living space in square meters (sqm).

Chapter #5

+++ estate/models/estate_property.py
@@ -14,7 +14,7 @@ class EstateProperty(models.Model):
     expected_price = fields.Float(string="Expected Price", required=True)
     facades = fields.Integer(string="Facades")
     garden = fields.Boolean(string="Garden")
-    garden_area = fields.Integer(string="Garden Area (sqm)")
+    garden_area = fields.Float(string="Garden Area (sqm)")
     garden_orientation = fields.Selection(
         string="Direction",
         selection=[
